couple of thoughts
